CrateDB vector: Unable to invoke `add_embeddings` without embeddings

+++ libs/langchain/langchain/vectorstores/cratedb/base.py
@@ -157,6 +157,8 @@ class CrateDBVectorSearch(PGVector):
             metadatas: List of metadatas associated with the texts.
             kwargs: vectorstore specific parameters
         """
+        if not embeddings:
+            return []
         self._init_models(embeddings[0])
         if self.pre_delete_collection:
             self.delete_collection()
